move to package and add py.typed
