+++ .gitignore
@@ -5,3 +5,4 @@
 /.eggs
 /.tox
 /build
+/dist
