+++ setup.cfg
@@ -19,12 +19,16 @@ classifiers =
     Programming Language :: Python :: Implementation :: PyPy
 
 [options]
-py_modules = ukkonen
+packages = ukkonen
 install_requires =
     cffi>=1
 python_requires = >=3.9
 setup_requires =
     cffi>=1
+zip_safe = false
+
+[options.package_data]
+* = py.typed
 
 [coverage:run]
 plugins = covdefaults
