+++ ukkonen/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-import _ukkonen
+import ukkonen._ukkonen
 
 _lib = _ukkonen.lib
