+++ ukkonen/_ukkonen.cpp
